refactor (Score) : Score 객체 삭제
